pyproject.toml
[tool.poetry]
name="robotframework-openapidriver"
version = "2.0.1"
description = "A library for contract-testing OpenAPI / Swagger APIs."
license = "Apache-2.0"
authors = ["Robin Mackaij"]
maintainers = ["Robin Mackaij"]
readme =  "./docs/README.md"
homepage = "https://github.com/MarketSquare/robotframework-openapidriver"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Testing :: Acceptance",
    "Framework :: Robot Framework",
]
packages = [
    { include = "openapidriver", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.8"
robotframework = ">=4"
robotframework-pythonlibcore = ">=3"
robotframework-datadriver = ">=1.5"
requests = "*"
prance = "*"
openapi-core = "*"
openapi-spec-validator = "*"

[tool.poetry.dev-dependencies]
black = "*"
isort = "*"
pylint = "*"
mypy = "*"
types-requests = "^2.25.6"
invoke = "*"
robotframework-stacktrace = "*"
robotframework-tidy = "*"
uvicorn = "*"
fastapi = "*"
coverage = {version = "^5.5", extras = ["toml"]}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
branch = true
parallel = true
source = ["src"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "@abstract"
]

[[tool.mypy.overrides]]
module = ["DataDriver.*", "prance.*", "robot.*", "openapi_core.*", "robotlibcore", "uvicorn"]
ignore_missing_imports = true

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
src_paths = [
    "src"
]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.pylint.'MESSAGES CONTROL']
disable = ["W1203"]

[tool.pylint.'FORMAT CHECKER']
max-line-length=120
